+++ packages/itwin/property-grid/src/property-grid-react.ts
@@ -6,7 +6,15 @@
 export * from "./property-grid-react/PropertyGridManager.js";
 export * from "./property-grid-react/PropertyGridComponent.js";
 export * from "./property-grid-react/api/PreferencesStorage.js";
-export * from "./property-grid-react/PropertyGridUiItemsProvider.js";
+export {
+  PropertyGridWidgetId,
+  PropertyGridWidgetProps,
+  createPropertyGrid,
+  // eslint-disable-next-line deprecation/deprecation
+  PropertyGridUiItemsProvider,
+  // eslint-disable-next-line deprecation/deprecation
+  PropertyGridUiItemsProviderProps,
+} from "./property-grid-react/PropertyGridUiItemsProvider.js";
 
 export * from "./property-grid-react/components/PropertyGrid.js";
 export * from "./property-grid-react/components/MultiElementPropertyGrid.js";