+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/categories-tree/UseCategoriesTree.tsx
@@ -5,11 +5,7 @@
 
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { assert } from "@itwin/core-bentley";
-import categorySvg from "@itwin/itwinui-icons/bis-category-3d.svg";
-import subcategorySvg from "@itwin/itwinui-icons/bis-category-subcategory.svg";
-import classSvg from "@itwin/itwinui-icons/bis-class.svg";
-import definitionContainerSvg from "@itwin/itwinui-icons/bis-definitions-container.svg";
-import elementSvg from "@itwin/itwinui-icons/bis-element.svg";
+import { Icon } from "@itwin/itwinui-react/bricks";
 import { createECSqlQueryExecutor } from "@itwin/presentation-core-interop";
 import { HierarchyFilteringPath, HierarchyNodeIdentifier } from "@itwin/presentation-hierarchies";
 import { EmptyTreeContent, FilterUnknownError, NoFilterMatches, TooManyFilterMatches } from "../common/components/EmptyTree.js";
@@ -48,7 +44,7 @@ interface UseCategoriesTreeResult {
     VisibilityTreeProps,
     "treeName" | "getHierarchyDefinition" | "getFilteredPaths" | "visibilityHandlerFactory" | "highlight" | "emptyTreeContent"
   >;
-  rendererProps: Required<Pick<VisibilityTreeRendererProps, "getIcon" | "getSublabel">>;
+  rendererProps: Required<Pick<VisibilityTreeRendererProps, "getDecorations" | "getSublabel">>;
 }
 
 function createVisibilityHandlerFactory(
@@ -178,7 +174,7 @@ export function useCategoriesTree({
       highlight: filter ? { text: filter } : undefined,
     },
     rendererProps: {
-      getIcon,
+      getDecorations: (node) => <CategoriesTreeIcon node={node} />,
       getSublabel,
     },
   };
@@ -251,6 +247,8 @@ async function getCategoriesFromPaths(paths: HierarchyFilteringPaths, idsCache: 
   }));
 }
 
+const categorySvg = new URL("@itwin/itwinui-icons/bis-category-3d.svg", import.meta.url).href;
+
 function getEmptyTreeContentComponent(filter?: string, error?: CategoriesTreeFilteringError, emptyTreeContent?: React.ReactNode) {
   if (error) {
     if (error === "tooManyFilterMatches") {
@@ -267,25 +265,35 @@ function getEmptyTreeContentComponent(filter?: string, error?: CategoriesTreeFil
   return <EmptyTreeContent icon={categorySvg} />;
 }
 
-function getIcon(node: PresentationHierarchyNode): string | undefined {
+const subcategorySvg = new URL("@itwin/itwinui-icons/bis-category-subcategory.svg", import.meta.url).href;
+const definitionContainerSvg = new URL("@itwin/itwinui-icons/bis-definitions-container.svg", import.meta.url).href;
+const classSvg = new URL("@itwin/itwinui-icons/bis-class.svg", import.meta.url).href;
+const elementSvg = new URL("@itwin/itwinui-icons/bis-element.svg", import.meta.url).href;
+
+/** @beta */
+export function CategoriesTreeIcon({ node }: { node: PresentationHierarchyNode }) {
   if (node.extendedData?.imageId === undefined) {
     return undefined;
   }
 
-  switch (node.extendedData.imageId) {
-    case "icon-layers":
-      return categorySvg;
-    case "icon-layers-isolate":
-      return subcategorySvg;
-    case "icon-definition-container":
-      return definitionContainerSvg;
-    case "icon-item":
-      return elementSvg;
-    case "icon-ec-class":
-      return classSvg;
-  }
+  const getIcon = () => {
+    switch (node.extendedData!.imageId) {
+      case "icon-layers":
+        return categorySvg;
+      case "icon-layers-isolate":
+        return subcategorySvg;
+      case "icon-definition-container":
+        return definitionContainerSvg;
+      case "icon-item":
+        return elementSvg;
+      case "icon-ec-class":
+        return classSvg;
+      default:
+        return undefined;
+    }
+  };
 
-  return undefined;
+  return <Icon href={getIcon()} />;
 }
 
 function getSublabel(node: PresentationHierarchyNode) {