+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/external-sources-tree/ExternalSourcesTree.tsx
@@ -3,10 +3,7 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import classSvg from "@itwin/itwinui-icons/bis-class.svg";
-import elementSvg from "@itwin/itwinui-icons/bis-element.svg";
-import documentSvg from "@itwin/itwinui-icons/document.svg";
-import ecSchemaSvg from "@itwin/itwinui-icons/selection-children.svg";
+import { Icon } from "@itwin/itwinui-react/bricks";
 import { EmptyTreeContent } from "../common/components/EmptyTree.js";
 import { Tree } from "../common/components/Tree.js";
 import { TreeRenderer } from "../common/components/TreeRenderer.js";
@@ -18,22 +15,24 @@ import type { BaseTreeRendererProps } from "../common/components/BaseTreeRendere
 import type { TreeProps } from "../common/components/Tree.js";
 /** @beta */
 export type ExternalSourcesTreeProps = Pick<TreeProps, "imodel" | "getSchemaContext" | "selectionStorage" | "selectionMode" | "emptyTreeContent"> &
-  Pick<BaseTreeRendererProps, "actions"> & {
+  Pick<BaseTreeRendererProps, "actions" | "getDecorations"> & {
     hierarchyLevelConfig?: {
       sizeLimit?: number;
     };
   };
 
 /** @beta */
-export function ExternalSourcesTree(props: ExternalSourcesTreeProps) {
+export function ExternalSourcesTree({ actions, getDecorations, selectionMode, ...rest }: ExternalSourcesTreeProps) {
   return (
     <Tree
       emptyTreeContent={<EmptyTreeContent icon={documentSvg} />}
-      {...props}
+      {...rest}
       treeName={ExternalSourcesTreeComponent.id}
       getHierarchyDefinition={getDefinitionsProvider}
-      selectionMode={props.selectionMode ?? "none"}
-      treeRenderer={(treeProps) => <TreeRenderer {...treeProps} getIcon={getIcon} />}
+      selectionMode={selectionMode ?? "none"}
+      treeRenderer={(treeProps) => (
+        <TreeRenderer {...treeProps} actions={actions} getDecorations={getDecorations ?? ((node) => <ExternalSourcesTreeIcon node={node} />)} />
+      )}
     />
   );
 }
@@ -42,21 +41,30 @@ const getDefinitionsProvider: TreeProps["getHierarchyDefinition"] = (props) => {
   return new ExternalSourcesTreeDefinition(props);
 };
 
-function getIcon(node: PresentationHierarchyNode): string | undefined {
+const classSvg = new URL("@itwin/itwinui-icons/bis-class.svg", import.meta.url).href;
+const elementSvg = new URL("@itwin/itwinui-icons/bis-element.svg", import.meta.url).href;
+const documentSvg = new URL("@itwin/itwinui-icons/document.svg", import.meta.url).href;
+const ecSchemaSvg = new URL("@itwin/itwinui-icons/selection-children.svg", import.meta.url).href;
+
+/** @beta */
+export function ExternalSourcesTreeIcon({ node }: { node: PresentationHierarchyNode }) {
   if (node.extendedData?.imageId === undefined) {
     return undefined;
   }
+  const getIcon = () => {
+    switch (node.extendedData!.imageId) {
+      case "icon-item":
+        return elementSvg;
+      case "icon-ec-class":
+        return classSvg;
+      case "icon-document":
+        return documentSvg;
+      case "icon-ec-schema":
+        return ecSchemaSvg;
+      default:
+        return undefined;
+    }
+  };
 
-  switch (node.extendedData.imageId) {
-    case "icon-item":
-      return elementSvg;
-    case "icon-ec-class":
-      return classSvg;
-    case "icon-document":
-      return documentSvg;
-    case "icon-ec-schema":
-      return ecSchemaSvg;
-  }
-
-  return undefined;
+  return <Icon href={getIcon()} />;
 }