+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/models-tree/ModelsTree.tsx
@@ -12,7 +12,7 @@ import type { UseModelsTreeProps } from "./UseModelsTree.js";
 import type { VisibilityTreeProps } from "../common/components/VisibilityTree.js";
 /** @beta */
 export type ModelsTreeProps = Pick<VisibilityTreeProps, "imodel" | "getSchemaContext" | "selectionStorage" | "selectionMode" | "emptyTreeContent"> &
-  Pick<VisibilityTreeRendererProps, "actions"> &
+  Pick<VisibilityTreeRendererProps, "actions" | "getDecorations"> &
   UseModelsTreeProps & {
     hierarchyLevelConfig?: {
       sizeLimit?: number;
@@ -32,6 +32,7 @@ export function ModelsTree({
   selectionPredicate,
   visibilityHandlerOverrides,
   getFilteredPaths,
+  getDecorations,
   onModelsFiltered,
   actions,
   emptyTreeContent,
@@ -55,7 +56,9 @@ export function ModelsTree({
       getSchemaContext={getSchemaContext}
       hierarchyLevelSizeLimit={hierarchyLevelConfig?.sizeLimit}
       selectionMode={selectionMode}
-      treeRenderer={(treeProps) => <VisibilityTreeRenderer {...treeProps} {...rendererProps} actions={actions} />}
+      treeRenderer={(treeProps) => (
+        <VisibilityTreeRenderer {...treeProps} {...rendererProps} actions={actions} getDecorations={getDecorations ?? rendererProps.getDecorations} />
+      )}
     />
   );
 }