+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/models-tree/UseModelsTree.tsx
@@ -4,11 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import categorySvg from "@itwin/itwinui-icons/bis-category-3d.svg";
-import classSvg from "@itwin/itwinui-icons/bis-class.svg";
-import elementSvg from "@itwin/itwinui-icons/bis-element.svg";
-import subjectSvg from "@itwin/itwinui-icons/bis-subject.svg";
-import modelSvg from "@itwin/itwinui-icons/model-cube.svg";
+import { Icon } from "@itwin/itwinui-react/bricks";
 import { createECSqlQueryExecutor } from "@itwin/presentation-core-interop";
 import { HierarchyNodeIdentifier, HierarchyNodeKey } from "@itwin/presentation-hierarchies";
 import {
@@ -38,7 +34,6 @@ import type { ClassGroupingHierarchyNode, ElementsGroupInfo, ModelsTreeHierarchy
 import type { ModelsTreeVisibilityHandlerOverrides } from "./internal/ModelsTreeVisibilityHandler.js";
 import type { VisibilityTreeProps } from "../common/components/VisibilityTree.js";
 import type { VisibilityTreeRendererProps } from "../common/components/VisibilityTreeRenderer.js";
-
 type ModelsTreeFilteringError = "tooManyFilterMatches" | "tooManyInstancesFocused" | "unknownFilterError" | "unknownInstanceFocusError";
 
 /** @beta */
@@ -65,7 +60,7 @@ interface UseModelsTreeResult {
     VisibilityTreeProps,
     "treeName" | "getHierarchyDefinition" | "getFilteredPaths" | "visibilityHandlerFactory" | "highlight" | "emptyTreeContent" | "selectionPredicate"
   >;
-  rendererProps: Required<Pick<VisibilityTreeRendererProps, "getIcon">>;
+  rendererProps: Required<Pick<VisibilityTreeRendererProps, "getDecorations">>;
 }
 
 /**
@@ -224,7 +219,7 @@ export function useModelsTree({
     },
     rendererProps: {
       // onDoubleClick,
-      getIcon,
+      getDecorations: (node) => <ModelsTreeIcon node={node} />,
     },
   };
 }
@@ -295,25 +290,36 @@ function InstanceFocusError({ error }: { error: ModelsTreeFilteringError }) {
   return <UnknownInstanceFocusError base={"modelsTree"} />;
 }
 
-function getIcon(node: PresentationHierarchyNode): string | undefined {
+const subjectSvg = new URL("@itwin/itwinui-icons/bis-subject.svg", import.meta.url).href;
+const classSvg = new URL("@itwin/itwinui-icons/bis-class.svg", import.meta.url).href;
+const modelSvg = new URL("@itwin/itwinui-icons/model-cube.svg", import.meta.url).href;
+const categorySvg = new URL("@itwin/itwinui-icons/bis-category-3d.svg", import.meta.url).href;
+const elementSvg = new URL("@itwin/itwinui-icons/bis-element.svg", import.meta.url).href;
+
+/** @beta */
+export function ModelsTreeIcon({ node }: { node: PresentationHierarchyNode }) {
   if (node.extendedData?.imageId === undefined) {
     return undefined;
   }
 
-  switch (node.extendedData.imageId) {
-    case "icon-layers":
-      return categorySvg;
-    case "icon-item":
-      return elementSvg;
-    case "icon-ec-class":
-      return classSvg;
-    case "icon-folder":
-      return subjectSvg;
-    case "icon-model":
-      return modelSvg;
-  }
+  const getIcon = () => {
+    switch (node.extendedData!.imageId) {
+      case "icon-layers":
+        return categorySvg;
+      case "icon-item":
+        return elementSvg;
+      case "icon-ec-class":
+        return classSvg;
+      case "icon-folder":
+        return subjectSvg;
+      case "icon-model":
+        return modelSvg;
+      default:
+        return undefined;
+    }
+  };
 
-  return undefined;
+  return <Icon href={getIcon()} />;
 }
 
 function createVisibilityHandlerFactory(